fix(LottoResults): 수익률 덧셈 data type long 으로 변경
수익률 덧셈 data type long 으로 변경
- 1등이 20억이라 초과 가능성 있음

+++ src/main/java/lotto/BuyAmount.java
@@ -27,7 +27,7 @@ public class BuyAmount {
         return amount * LOTTO_TICKET_PRICE;
     }
 
-    public double getProfitRate(int totalReward) {
+    public double getProfitRate(long totalReward) {
         double basePrice = getPrice();
         return (totalReward - basePrice) / basePrice;
     }
